+++ tests/utils.js
@@ -2,8 +2,8 @@
 const assert = require('assert');
 
 module.exports = {
-    async assertThrowsAssertionAsync (fn, expected) {
-        return this.assertThrowsAsync (fn, `AssertionError [ERR_ASSERTION]: ${expected}`);
+    async assertThrowsAssertionAsync (fn, expectedMsg) {
+        return this.assertThrowsAsync (fn, `AssertionError [ERR_ASSERTION]: ${expectedMsg}`);
     },
     async assertThrowsAsync (fn, expected) {
         let f = () => {};