+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 ==================
 
   * Browser.assert.global
-  * Added Custom Errors
+  * Added custom errors
+  * Assertion error improvements
   * Dependencies update
 
 0.4.4 / 2018-03-23